+++ service/src/main/scala/org/apache/celeborn/server/common/HttpService.scala
@@ -23,11 +23,11 @@ import scala.collection.JavaConverters._
 
 import org.apache.celeborn.common.CelebornConf
 import org.apache.celeborn.common.internal.Logging
-import org.apache.celeborn.server.common.http.{HttpRequestHandler, HttpServer, HttpServerInitializer}
+import org.apache.celeborn.server.common.http.HttpServer
+import org.apache.celeborn.server.common.http.api.ApiRootResource
 import org.apache.celeborn.server.common.service.config.ConfigLevel
 
 abstract class HttpService extends Service with Logging {
-
   private var httpServer: HttpServer = _
 
   def getConf: String = {
@@ -173,18 +173,19 @@ abstract class HttpService extends Service with Logging {
   def getWorkerEventInfo(): String = throw new UnsupportedOperationException()
 
   def startHttpServer(): Unit = {
-    val handlers =
-      if (metricsSystem.running) {
-        new HttpRequestHandler(this, metricsSystem.getServletHandlers)
-      } else {
-        new HttpRequestHandler(this, null)
-      }
-    httpServer = new HttpServer(
+    httpServer = HttpServer(
       serviceName,
       httpHost(),
       httpPort(),
-      new HttpServerInitializer(handlers))
+      999)
     httpServer.start()
+    startInternal()
+    // block until the HTTP server is started, otherwise, we may get
+    // the wrong HTTP server port -1
+    while (httpServer.getState != "STARTED") {
+      logInfo(s"Waiting for $serviceName's HTTP server getting started")
+      Thread.sleep(1000)
+    }
   }
 
   private def httpHost(): String = {
@@ -203,6 +204,17 @@ abstract class HttpService extends Service with Logging {
       case Service.WORKER =>
         conf.workerHttpPort
     }
+  }
+
+  def connectionUrl: String = {
+    httpServer.getServerUri
+  }
+
+  protected def startInternal(): Unit = {
+    httpServer.addHandler(ApiRootResource.getServletHandler(this))
+    httpServer.addStaticHandler("org/apache/celeborn/swagger", "/swagger")
+    httpServer.addRedirectHandler("/help", "/swagger")
+    httpServer.addRedirectHandler("/docs", "/swagger")
   }
 
   override def initialize(): Unit = {
